requirements.txt
pandas
matplotlib
seaborn
numpypandas>=1.5
matplotlib>=3.5
seaborn>=0.12
